add dry-run mode

+++ src/commands.py
@@ -1,11 +1,17 @@
 import os
 import shutil
 
+from utils import Mode
 
-def remove_file(fpath):
-    os.remove(fpath)
-    print("-> {} deleted.".format(fpath))
-    return {"path": fpath}
+
+def remove_file(fpath, exec_mode):
+    try:
+        if exec_mode not in (Mode.DryRun, Mode.Both):
+            os.remove(fpath)
+        print("-> {} deleted.".format(fpath))
+        return {"path": fpath}
+    except OSError as e:
+        print("-> Error: {}".format(e))
 
 
 def apply_removed(all_files, changes):
@@ -13,10 +19,14 @@ def apply_removed(all_files, changes):
         all_files.pop(change["path"])
 
 
-def rename_file(fpath, new_fpath):
-    os.rename(fpath, new_fpath)
-    print("-> {} renamed to {}".format(fpath, new_fpath))
-    return {"old_path": fpath, "new_path": new_fpath}
+def rename_file(fpath, new_fpath, exec_mode):
+    try:
+        if exec_mode not in (Mode.DryRun, Mode.Both):
+            os.rename(fpath, new_fpath)
+        print("-> {} renamed to {}".format(fpath, new_fpath))
+        return {"old_path": fpath, "new_path": new_fpath}
+    except OSError as e:
+        print("-> Error: {}".format(e))
 
 
 def apply_renamed(all_files, changes):
@@ -28,10 +38,14 @@ def apply_renamed(all_files, changes):
         all_files[new_path] = entry
 
 
-def chmod_file(fpath, mode):
-    os.chmod(fpath, mode)
-    print("-> Changed mode of {} to {:o}".format(fpath, mode))
-    return {"path": fpath, "mode": mode}
+def chmod_file(fpath, mode, exec_mode):
+    try:
+        if exec_mode not in (Mode.DryRun, Mode.Both):
+            os.chmod(fpath, mode)
+        print("-> Changed mode of {} to {:o}".format(fpath, mode))
+        return {"path": fpath, "mode": mode}
+    except OSError as e:
+        print("-> Error: {}".format(e))
 
 
 def apply_chmod(all_files, changes):
@@ -40,10 +54,14 @@ def apply_chmod(all_files, changes):
         all_files[path].mode = change["mode"]
 
 
-def copy_file(src, dest, new_root, new_kind):
-    shutil.copy(src, dest)
-    print("-> File moved from {} to {}".format(src, dest))
-    return {"old_path": src, "new_path": dest, "new_root": new_root, "new_kind": new_kind}
+def copy_file(src, dest, new_root, new_kind, exec_mode):
+    try:
+        if exec_mode not in (Mode.DryRun, Mode.Both):
+            shutil.copy(src, dest)
+        print("-> File moved from {} to {}".format(src, dest))
+        return {"old_path": src, "new_path": dest, "new_root": new_root, "new_kind": new_kind}
+    except OSError as e:
+        print("-> Error: {}".format(e))
 
 
 def apply_copied(all_files, changes):
